1. Ported OTP functionality 2. Ported MFA functionality for Super Admins 3. Ported more changes and updates 4. Use Sentry for RR

+++ packages/features/accounts/src/server/personal-account-actions.server.ts
@@ -5,6 +5,7 @@ import { SupabaseClient } from '@supabase/supabase-js';
 import process from 'node:process';
 import { z } from 'zod';
 
+import { createOtpApi } from '@kit/otp';
 import { Database } from '@kit/supabase/database';
 import { requireUser } from '@kit/supabase/require-user';
 import { getSupabaseServerAdminClient } from '@kit/supabase/server-admin-client';
@@ -15,8 +16,10 @@ const emailSettings = getEmailSettingsFromEnvironment();
 
 export const deletePersonalAccountAction = async ({
   client,
+  otp,
 }: {
   client: SupabaseClient<Database>;
+  otp: string;
 }) => {
   const auth = await requireUser(client);
 
@@ -29,8 +32,23 @@ export const deletePersonalAccountAction = async ({
   // create a new instance of the personal accounts service
   const service = createDeletePersonalAccountService();
 
-  // sign out the user before deleting their account
-  await client.auth.signOut();
+  // verify OTP
+  const otpApi = createOtpApi(client);
+
+  const result = await otpApi.verifyToken({
+    purpose: 'delete-personal-account',
+    userId: user.id,
+    token: otp,
+  });
+
+  if (!result.valid) {
+    throw new Error('Invalid OTP');
+  }
+
+  // validate the user ID matches the nonce's user ID
+  if (result.user_id !== user.id) {
+    throw new Error('Nonce mismatch');
+  }
 
   // delete the user's account and cancel all subscriptions
   await service.deletePersonalAccount({
@@ -39,6 +57,9 @@ export const deletePersonalAccountAction = async ({
     userEmail: user.email ?? null,
     emailSettings,
   });
+
+  // sign out the user before deleting their account
+  await client.auth.signOut();
 
   // redirect to the home page
   return redirectDocument('/');
